move promptShouldAutoStartDesignTime to main, update promptShouldEnableLocalManagedIdentityAuth to be consistent
